prepare: not symlink self path

+++ tools/prepare.sh
@@ -100,9 +100,12 @@ then
     if [ -f $m/Frameworks/BUILD_PATH ]
     then
       BUILD_PATH=`cat $m/Frameworks/BUILD_PATH`
-      sudo mkdir -p $(dirname $BUILD_PATH)
-      sudo rm -f $BUILD_PATH
-      sudo ln -s $(pwd)/$m $BUILD_PATH
+      if [ ! -d $BUILD_PATH ]
+      then
+        sudo mkdir -p $(dirname $BUILD_PATH)
+        sudo rm -f $BUILD_PATH
+        sudo ln -s $(pwd)/$m $BUILD_PATH
+      fi
     fi
   done
